Fix signed comparisons in utils/vrad.

+++ sp/src/utils/vrad/vismat.cpp
@@ -207,7 +207,7 @@ Sets vis bits for all patches in the face
 */
 void TestPatchToFace (unsigned patchnum, int facenum, int head, transfer_t *transfers, CTransferMaker &transferMaker, int iThread )
 {
-	if( faceParents.Element( facenum ) == g_Patches.InvalidIndex() || patchnum == g_Patches.InvalidIndex() )
+	if( faceParents.Element( facenum ) == g_Patches.InvalidIndex() || patchnum == (unsigned) g_Patches.InvalidIndex() )
 		return;
 
 	CPatch	*patch = &g_Patches.Element( patchnum );
